+++ dev/back-end/BackLaizeTech/BackLaizeTech/src/main/kotlin/laize_tech/back/dto/CaracteristicaDTO.kt
@@ -2,5 +2,5 @@ package laize_tech.back.dto
 
 data class CaracteristicaDTO(
     val nomeCaracteristica: String,
-    val idTipoCaracteristica: Long
+    val idTipoCaracteristica: Int
 )
