+++ dev/back-end/BackLaizeTech/BackLaizeTech/src/main/kotlin/laize_tech/back/repository/TipoCaracteristicaRepository.kt
@@ -3,5 +3,5 @@ package laize_tech.back.repository
 import laize_tech.back.entity.TipoCaracteristica
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface TipoCaracteristicaRepository  : JpaRepository<TipoCaracteristica, Long> {
+interface TipoCaracteristicaRepository  : JpaRepository<TipoCaracteristica, Int> {
 }
